:sparkles:  Feat: API 함수 수정 및 HomePage 적용
Closes #3
