+++ src/App.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { GlobalStyle } from '@style/GlobalStyle'
+import HomePage from '@pages/HomePage'
 
 const App = () => {
   return (
@@ -11,7 +12,7 @@ const App = () => {
           <Route></Route>
         </Routes>
       </BrowserRouter>
-      <button>hi</button>
+      <HomePage />
     </>
   )
 }
