+++ src/pages/index.jsx
@@ -1,0 +1,1 @@
+export { default as HomePage } from './HomePage'
